requirements.txt
aniso8601==1.1.0
argon2-cffi==16.2.0
cffi==1.8.3
click==6.6
coverage==4.2
Flask==0.11.1
Flask-HTTPAuth==3.2.3
Flask-RESTful==0.3.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
peewee==2.8.1
pycparser==2.18
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
Werkzeug==0.11.10
